fix overflow bug in PriceFilterOption

+++ app/src/main/java/com/bamilo/android/appmodule/bamiloapp/utils/catalog/filters/FilterPriceFragment.java
@@ -26,7 +26,7 @@ import com.bamilo.android.appmodule.bamiloapp.view.fragments.FilterMainFragment;
  * @author sergiopereira
  *
  */
-public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBarChangeListener<Integer> {
+public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBarChangeListener<Long> {
 
     private static final String TAG = FilterPriceFragment.class.getSimpleName();
 
@@ -34,17 +34,17 @@ public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBa
     private XeiEditText mMinValueTxt;
     private XeiEditText mMaxValueTxt;
 
-    private int mMin;
+    private long mMin;
 
-    private int mMax;
+    private long mMax;
 
-    private int mCurrMinValue ,mMinRang;
+    private long mCurrMinValue ,mMinRang;
 
-    private int mCurrMaxValue, mMaxRang;
+    private long mCurrMaxValue, mMaxRang;
 
-    private RangeSeekBar<Integer> mRangeBar;
+    private RangeSeekBar<Long> mRangeBar;
 
-    private int mInterval;
+    private long mInterval;
 
     private CatalogPriceFilter mFilter;
 
@@ -95,7 +95,7 @@ public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBa
                 if (mymax<=mMaxRang &&Integer.MAX_VALUE>mymax) {
 
                     setIntervalText(mCurrMaxValue,(int) mymax.longValue());
-                    mRangeBar.setSelectedMaxValue(getMaxIntervalValue((int) mymax.longValue()));
+                    mRangeBar.setSelectedMaxValue(getMaxIntervalValue(mymax.longValue()));
                     mCurrMaxValue =mymax.intValue() ;
                     mFilter.getOption().setRangeMin(mCurrMinValue);
                     mFilter.getOption().setRangeMax(mCurrMaxValue);
@@ -169,7 +169,7 @@ public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBa
     }
 
     @Override
-    public void onRangeSeekBarValuesChanged(RangeSeekBar<?> bar, Integer minValue, Integer maxValue) {
+    public void onRangeSeekBarValuesChanged(RangeSeekBar<?> bar, Long minValue, Long maxValue) {
         mCurrMinValue = getMinRealValue(minValue);
         mCurrMaxValue = getMaxRealValue(maxValue);
         setIntervalText(mCurrMinValue, mCurrMaxValue);
@@ -218,7 +218,7 @@ public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBa
      * @return max interval value
      * @author sergiopereira
      */
-    private int getMaxIntervalValue(int max){
+    private long getMaxIntervalValue(long max){
         return (mInterval != 0) ? max / mInterval : max;   
     }
 
@@ -228,7 +228,7 @@ public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBa
      * @return min interval value
      * @author sergiopereira
      */
-    private int getMinIntervalValue(int min){
+    private long getMinIntervalValue(long min){
         return (mInterval != 0) ? min / mInterval : min;   
     }
     
@@ -238,7 +238,7 @@ public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBa
      * @return min real value
      * @author sergiopereira
      */
-    private int getMinRealValue(int min){
+    private long getMinRealValue(long min){
         return (mInterval != 0) ? min * mInterval : min;   
     }
     
@@ -248,7 +248,7 @@ public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBa
      * @return min real value
      * @author sergiopereira
      */
-    private int getMaxRealValue(int max){
+    private long getMaxRealValue(long max){
         return (mInterval != 0) ? max * mInterval : max;   
     }
 
@@ -257,7 +257,7 @@ public class FilterPriceFragment extends FilterFragment implements OnRangeSeekBa
         processOnClickClean();
     }
 
-    protected void setIntervalText(@IntRange(from = 0) int minValue, int maxValue) {
+    protected void setIntervalText(@IntRange(from = 0) long minValue, long maxValue) {
         mRangeValues.setText(CurrencyFormatter.formatCurrencyPattern(minValue + " - " + maxValue));
     }
